+++ src/app/components/card/card.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 
-import { CardService } from './card.service';
+import { CardService } from '../../services/card.service';
 
 describe('CardService', () => {
   beforeEach(() => TestBed.configureTestingModule({}));
