+++ src/app/interfaces/card.model.ts
