+++ src/app/pages/payload/transaction-response-payload.model.ts
@@ -1,0 +1,4 @@
+export interface TransactionResponsePayload {
+    success: boolean,
+    status: string
+}
